Nearly complete refactoring.
