+++ src/hydra/utils/apply.py
@@ -1,307 +1,11 @@
 """Module responsible for applying heightmaps, previews and modifiers."""
 
-import bpy, bpy.types, bpy_extras
+import bpy
 import numpy as np
 from Hydra import common
 from Hydra.sim import heightmap
-from Hydra.utils import texture, nav
-import moderngl as mgl
-import math, re
-
-def minimize_node(node):
-	"""Hides unused inputs and minimizes the specified node.
-	
-	:param node: Node to minimize."""
-	node.hide = True
-	for n in node.inputs:
-		n.hide = True
-	for n in node.outputs:
-		n.hide = True
-
-def setup_vector_node(nodes, node: bpy.types.ShaderNode)->bpy.types.ShaderNode:
-	"""Creates a Z Normal node and connects it to the specified node.
-	
-	:param nodes: Node graph.
-	:param node: Node to connect to.
-	:type node: :class:`bpy.types.ShaderNode`
-	:return: Created node.
-	:rtype: :class:`bpy.types.ShaderNode`"""
-	norm = nodes.nodes.new("ShaderNodeNormal")
-	norm.name = "HYD_norm"
-	norm.label = "Z Normal"
-	nodes.links.new(node.inputs["Normal"], norm.outputs["Normal"])
-	minimize_node(norm)
-	return norm
-	
-def setup_image_node(nodes, name:str, imageSrc:str)->tuple[bpy.types.ShaderNode, bpy.types.ShaderNode]:
-	"""Creates an Image Texture node and connects it to generated coordinates.
-	
-	:param nodes: Node graph.
-	:param name: Image node title.
-	:type name: :class:`str`
-	:param imageSrc: Image source name.
-	:type imageSrc: :class:`str`
-	:return: Created nodes.
-	:rtype: :class:`tuple[bpy.types.ShaderNode, bpy.types.ShaderNode]`"""
-	img = nodes.nodes.new("ShaderNodeTexImage")
-	img.name = name
-	img.label = name
-	img.image = imageSrc
-	img.extension = 'EXTEND'
-	img.interpolation = 'Cubic'
-	coords = nodes.nodes.new("ShaderNodeTexCoord")
-	nodes.links.new(img.inputs["Vector"], coords.outputs["Generated"])
-	minimize_node(coords)
-	return (img, coords)
-	
-def stagger_nodes(baseNode:bpy.types.ShaderNode, *args, forwards:bool=False):
-	"""Spaces and shifts specified nodes around.
-	
-	:param baseNode: Rightmost node.
-	:type baseNode: :class:`bpy.types.ShaderNode`
-	:param args: Node arguments to be shifted.
-	:type args: :class:`bpy.types.ShaderNode`
-	:param forwards: Shift direction. `True` shifts `baseNode` forward.
-	:type forwards: :class:`bool`"""
-	x = baseNode.location[0]
-	baseY = baseNode.location[1] - baseNode.height
-	if forwards:
-		for layer in args[::-1]:
-			maxwidth = 0
-			y = baseY
-			for node in layer:
-				node.location[0] = x
-				maxwidth = max(maxwidth, node.width)
-				node.location[1] = y
-				y -= node.height + 10
-			x += maxwidth + 20
-		baseNode.location[0] = x
-	else:
-		for layer in args:
-			maxwidth = 0
-			y = baseY
-			for node in layer:
-				node.location[0] = x - node.width - 20
-				maxwidth = max(maxwidth, node.width)
-				node.location[1] = y
-				y -= node.height + 10
-			x -= maxwidth + 20
-			
-def make_bsdf(nodes)->bpy.types.ShaderNode:
-	"""Creates a Principled BSDF node.
-	
-	:param nodes: Node graph.
-	:return: Created node.
-	:rtype: :class:`bpy.types.ShaderNode`"""
-	ret = nodes.nodes.new("ShaderNodeBsdfPrincipled")
-	ret.inputs["Roughness"].default_value = 0.8
-	
-	if "Specular IOR Level" in ret.inputs:	# Blender 4.0
-		ret.inputs["Specular IOR Level"].default_value = 0.2
-	elif "Specular":	# Older versions
-		ret.inputs["Specular"].default_value = 0.2
-
-	return ret
-
-def get_or_make_output_node(nodes):
-	"""Finds or creates an Output node.
-	
-	:param nodes: Node graph.
-	:return: Created node.
-	:rtype: :class:`bpy.types.ShaderNode`"""
-	out = nodes.get_output_node('ALL')
-	if out is None:
-		out = nodes.get_output_node('CYCLES')
-
-	if out is None:
-		out = nodes.get_output_node('EEVEE')
-
-	if out is None:
-		out = nodes.nodes.new("ShaderNodeOutputMaterial")
-	
-	return out
-
-# --------------------------------------------------
-
-H_NAME_BUMP = "Hydra Bump"
-"""Bump map node name."""
-
-def add_bump(obj: bpy.types.Object, src: mgl.Texture):
-	"""Adds a Bump map for the given texture in the specified object's material.
-	Creates a material if needed, otherwise selects the first material slot.
-
-	:param obj: Object to add to.
-	:type obj: :class:`bpy.types.Object`
-	:param src: Source texture to add.
-	:type src: :class:`moderngl.Texture`"""
-	data = common.data
-	img = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
-
-	mats = obj.material_slots
-	if len(mats) == 0:
-		name = "HYD_" + obj.name
-		if name in bpy.data.materials:
-			mat = bpy.data.materials[name]
-		else:
-			mat = bpy.data.materials.new(name=name)
-		obj.data.materials.append(mat)
-	else:
-		mat = mats[0].material
-	
-	if not mat.use_nodes:
-		mat.use_nodes = True
-	nodes = mat.node_tree
-	
-	if H_NAME_BUMP in nodes.nodes: #already updated
-		nodes.nodes[H_NAME_BUMP].image = img
-		return
-	
-	out = get_or_make_output_node(nodes)
-		
-	if not out.inputs["Surface"].is_linked:
-		shader = make_bsdf(nodes)
-		nodes.links.new(out.inputs["Surface"], shader.outputs["BSDF"])
-	else:
-		shader = out.inputs["Surface"].links[0].from_node
-	if "Normal" not in shader.inputs:
-		data.error += ["Material doesn't have normals input.\n"]
-		return
-		
-	if shader.inputs["Normal"].is_linked:
-		bump = shader.inputs["Normal"].links[0].from_node
-		if bump.bl_idname != "ShaderNodeBump":
-			data.error += ["Material already implements a different bump map.\n"]
-			return
-	else:
-		bump = nodes.nodes.new("ShaderNodeBump")
-		nodes.links.new(shader.inputs["Normal"], bump.outputs["Normal"])
-	
-	if bump.inputs["Height"].is_linked:
-		data.error += ["Material already implements a different bump map.\n"]
-		return
-	
-	bump.inputs["Distance"].default_value = 0.2
-	
-	imgNode,coords = setup_image_node(nodes, "Bumpmap", img)
-	nodes.links.new(bump.inputs["Height"], imgNode.outputs["Color"])
-	minimize_node(imgNode)
-	stagger_nodes(shader, [bump], [imgNode], [coords])
-
-H_NAME_DISP = "Hydra Displacement"
-"""Displacement map node name."""
-
-def add_displacement(obj: bpy.types.Object, src: mgl.Texture):
-	"""Adds a Displacement map for the given texture in the specified object's material.
-	Creates a material if needed, otherwise selects the first material slot.
-
-	:param obj: Object to add to.
-	:type obj: :class:`bpy.types.Object`
-	:param src: Source texture to add.
-	:type src: :class:`moderngl.Texture`"""
-	data = common.data
-	img = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
-
-	mats = obj.material_slots
-	if len(mats) == 0:
-		name = "HYD_" + obj.name
-		if name in bpy.data.materials:
-			mat = bpy.data.materials[name]
-		else:
-			mat = bpy.data.materials.new(name=name)
-			mat.cycles.displacement_method = "DISPLACEMENT"
-		obj.data.materials.append(mat)
-	else:
-		mat = mats[0].material
-	
-	if not mat.use_nodes:
-		mat.use_nodes = True
-	nodes = mat.node_tree
-	
-	if H_NAME_DISP in nodes.nodes:	#Image node already exists
-		nodes.nodes[H_NAME_DISP].image = img
-		return
-	
-	out = get_or_make_output_node(nodes)
-
-	if not out.inputs["Displacement"].is_linked:
-		disp = nodes.nodes.new("ShaderNodeDisplacement")
-		disp.inputs["Midlevel"].default_value = 0
-		disp.inputs["Scale"].default_value = 1
-		nodes.links.new(out.inputs["Displacement"], disp.outputs["Displacement"])
-		norm = setup_vector_node(nodes, disp)
-	else:
-		disp = out.inputs["Displacement"].links[0].from_node
-		if disp.bl_idname != "ShaderNodeDisplacement":
-			data.error += ["Material already implements a different displacement map."]
-			return
-	
-	if disp.inputs["Height"].is_linked:
-		data.error += ["Material already implements a different displacement map."]
-		return
-	
-	imgNode, coords = setup_image_node(nodes, H_NAME_DISP, img)
-	nodes.links.new(disp.inputs["Height"], imgNode.outputs["Color"])
-	minimize_node(imgNode)
-	
-	stagger_nodes(out, [disp], [imgNode, norm], [coords], forwards=True)
-
-def add_modifier(obj: bpy.types.Object, src: mgl.Texture):
-	"""Adds a Displace modifier to the specified object.
-
-	:param obj: Object to add to.
-	:type obj: :class:`bpy.types.Object`
-	:param src: Source texture to add.
-	:type src: :class:`moderngl.Texture`"""
-	data = common.data
-	img = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
-
-	if len(obj.modifiers) == 0:
-		mod = obj.modifiers.new("HYD_" + obj.name, "DISPLACE")
-	elif obj.modifiers[-1].name.startswith("HYD_"):
-		if obj.modifiers[-1].type == "DISPLACE":
-			mod = obj.modifiers[-1]
-		else:
-			obj.modifiers.remove(obj.modifiers[-1])
-			mod = obj.modifiers.new("HYD_" + obj.name, "DISPLACE")
-	else:
-		mod = obj.modifiers[-1]
-	
-	if mod.name in bpy.data.textures:
-		txt = bpy.data.textures[mod.name]
-		data.info += ["Updated existing displacement."]
-	else:
-		txt = bpy.data.textures.new(mod.name, "IMAGE")
-		data.info += ["Created new displacement."]
-
-	txt.image = img
-	mod.texture = txt
-	mod.direction = 'Z'
-	mod.texture_coords = "OBJECT"
-	mod.strength = 1
-	mod.mid_level = 0
-	
-	coll = obj.users_collection[0]	#always exists
-	emptyName = "HYD_" + obj.name + "_Guide"
-	if emptyName in bpy.data.objects:
-		empty = bpy.data.objects[emptyName]
-	else:
-		empty = bpy.data.objects.new(emptyName, None)
-		coll.objects.link(empty)
-	empty.parent = obj
-	
-	ar = np.array(obj.bound_box)
-	# bounding box center
-	cx = (ar[4][0] + ar[0][0]) * 0.5
-	cy = (ar[2][1] + ar[0][1]) * 0.5
-	cz = (ar[1][2] + ar[0][2]) * 0.5
-	# bounding box size
-	sx = 0.5*(ar[4][0] - ar[0][0])
-	sy = 0.5*(ar[2][1] - ar[0][1])
-	# Z scale is handled by texture
-
-	empty.location = (cx,cy,cz)
-	empty.scale = (sx,sy,1)
-	mod.texture_coords_object = empty
+from Hydra.utils import texture, nav, nodes
+import math
 
 # -------------------------------------------------- Previews
 
@@ -335,6 +39,9 @@ def add_preview(target: bpy.types.Object|bpy.types.Image):
 	data = common.data
 	hyd = target.hydra_erosion
 
+	if not common.data.has_map(hyd.map_result):
+		return
+
 	if isinstance(target, bpy.types.Image):
 		img = texture.write_image(P_VIEW_NAME, data.maps[hyd.map_result].texture)
 		nav.goto_image(img)
@@ -358,7 +65,7 @@ def add_preview(target: bpy.types.Object|bpy.types.Image):
 			common.data.add_message("Created preview modifier.")
 		
 		img = heightmap.get_displacement(target, P_IMG_NAME)
-		mod.node_group = get_or_make_displace_group(P_GEO_NAME, img)
+		mod.node_group = nodes.get_or_make_displace_group(P_GEO_NAME, img)
 
 		common.data.lastPreview = target.name
 
@@ -396,6 +103,188 @@ def remove_preview():
 
 	data.lastPreview = ""
 
+# -------------------------------------------------- Shaders
+
+H_NAME_BUMP = "Hydra Bump"
+"""Bump map node name."""
+
+def add_bump(obj: bpy.types.Object, img: bpy.types.Image):
+	"""Adds a Bump map for the given texture in the specified object's material.
+	Creates a material if needed, otherwise selects the first material slot.
+
+	:param obj: Object to add to.
+	:type obj: :class:`bpy.types.Object`
+	:param src: Source texture to add.
+	:type src: :class:`moderngl.Texture`"""
+	data = common.data
+
+	mats = obj.material_slots
+	if len(mats) == 0:
+		name = "HYD_" + obj.name
+		if name in bpy.data.materials:
+			mat = bpy.data.materials[name]
+		else:
+			mat = bpy.data.materials.new(name=name)
+		obj.data.materials.append(mat)
+	else:
+		mat = mats[0].material
+	
+	if not mat.use_nodes:
+		mat.use_nodes = True
+	tree = mat.node_tree
+	
+	if H_NAME_BUMP in tree.nodes: #already updated
+		tree.nodes[H_NAME_BUMP].image = img
+		return
+	
+	out = nodes.get_or_make_output_node(tree)
+		
+	if not out.inputs["Surface"].is_linked:
+		shader = nodes.make_bsdf(tree)
+		tree.links.new(out.inputs["Surface"], shader.outputs["BSDF"])
+	else:
+		shader = out.inputs["Surface"].links[0].from_node
+	if "Normal" not in shader.inputs:
+		data.error += ["Material doesn't have normals input.\n"]
+		return
+		
+	if shader.inputs["Normal"].is_linked:
+		bump = shader.inputs["Normal"].links[0].from_node
+		if bump.bl_idname != "ShaderNodeBump":
+			data.error += ["Material already implements a different bump map.\n"]
+			return
+	else:
+		bump = tree.nodes.new("ShaderNodeBump")
+		tree.links.new(shader.inputs["Normal"], bump.outputs["Normal"])
+	
+	if bump.inputs["Height"].is_linked:
+		data.error += ["Material already implements a different bump map.\n"]
+		return
+	
+	bump.inputs["Distance"].default_value = 0.2
+	
+	imgNode,coords = nodes.setup_image_node(tree, "Bumpmap", img)
+	tree.links.new(bump.inputs["Height"], imgNode.outputs["Color"])
+	nodes.minimize_node(imgNode)
+	nodes.stagger_nodes(shader, [bump], [imgNode], [coords])
+
+H_NAME_DISP = "Hydra Displacement"
+"""Displacement map node name."""
+
+def add_displacement(obj: bpy.types.Object, img: bpy.types.Image):
+	"""Adds a Displacement map for the given texture in the specified object's material.
+	Creates a material if needed, otherwise selects the first material slot.
+
+	:param obj: Object to add to.
+	:type obj: :class:`bpy.types.Object`
+	:param src: Source texture to add.
+	:type src: :class:`moderngl.Texture`"""
+	data = common.data
+
+	mats = obj.material_slots
+	if len(mats) == 0:
+		name = "HYD_" + obj.name
+		if name in bpy.data.materials:
+			mat = bpy.data.materials[name]
+		else:
+			mat = bpy.data.materials.new(name=name)
+			mat.cycles.displacement_method = "DISPLACEMENT"
+		obj.data.materials.append(mat)
+	else:
+		mat = mats[0].material
+	
+	if not mat.use_nodes:
+		mat.use_nodes = True
+	tree = mat.node_tree
+	
+	if H_NAME_DISP in tree.nodes:	#Image node already exists
+		tree.nodes[H_NAME_DISP].image = img
+		return
+	
+	out = nodes.get_or_make_output_node(tree)
+
+	if not out.inputs["Displacement"].is_linked:
+		disp = tree.nodes.new("ShaderNodeDisplacement")
+		disp.inputs["Midlevel"].default_value = 0
+		disp.inputs["Scale"].default_value = 1
+		tree.links.new(out.inputs["Displacement"], disp.outputs["Displacement"])
+		norm = nodes.setup_vector_node(tree, disp)
+	else:
+		disp = out.inputs["Displacement"].links[0].from_node
+		if disp.bl_idname != "ShaderNodeDisplacement":
+			data.error += ["Material already implements a different displacement map."]
+			return
+	
+	if disp.inputs["Height"].is_linked:
+		data.error += ["Material already implements a different displacement map."]
+		return
+	
+	imgNode, coords = nodes.setup_image_node(tree, H_NAME_DISP, img)
+	tree.links.new(disp.inputs["Height"], imgNode.outputs["Color"])
+	nodes.minimize_node(imgNode)
+	
+	nodes.stagger_nodes(out, [disp], [imgNode, norm], [coords], forwards=True)
+
+# -------------------------------------------------- Modifier
+
+def add_modifier(obj: bpy.types.Object, img: bpy.types.Image):
+	"""Adds a Displace modifier to the specified object.
+
+	:param obj: Object to add to.
+	:type obj: :class:`bpy.types.Object`
+	:param src: Source texture to add.
+	:type src: :class:`moderngl.Texture`"""
+	data = common.data
+
+	if len(obj.modifiers) == 0:
+		mod = obj.modifiers.new("HYD_" + obj.name, "DISPLACE")
+	elif obj.modifiers[-1].name.startswith("HYD_"):
+		if obj.modifiers[-1].type == "DISPLACE":
+			mod = obj.modifiers[-1]
+		else:
+			obj.modifiers.remove(obj.modifiers[-1])
+			mod = obj.modifiers.new("HYD_" + obj.name, "DISPLACE")
+	else:
+		mod = obj.modifiers[-1]
+	
+	if mod.name in bpy.data.textures:
+		txt = bpy.data.textures[mod.name]
+		data.add_message("Updated existing displacement.")
+	else:
+		txt = bpy.data.textures.new(mod.name, "IMAGE")
+		data.add_message("Created new displacement.")
+
+	txt.image = img
+	mod.texture = txt
+	mod.direction = 'Z'
+	mod.texture_coords = "OBJECT"
+	mod.strength = 1
+	mod.mid_level = 0
+	
+	coll = obj.users_collection[0]	#always exists
+	emptyName = "HYD_" + obj.name + "_Guide"
+	if emptyName in bpy.data.objects:
+		empty = bpy.data.objects[emptyName]
+	else:
+		empty = bpy.data.objects.new(emptyName, None)
+		coll.objects.link(empty)
+	empty.parent = obj
+	
+	ar = np.array(obj.bound_box)
+	# bounding box center
+	cx = (ar[4][0] + ar[0][0]) * 0.5
+	cy = (ar[2][1] + ar[0][1]) * 0.5
+	cz = (ar[1][2] + ar[0][2]) * 0.5
+	# bounding box size
+	sx = 0.5*(ar[4][0] - ar[0][0])
+	sy = 0.5*(ar[2][1] - ar[0][1])
+	# Z scale is handled by texture
+
+	empty.location = (cx,cy,cz)
+	empty.scale = (sx,sy,1)
+	mod.texture_coords_object = empty
+
+# -------------------------------------------------- Landscape
 
 P_LAND_TEMP_NAME = "HYD_TEMP_DISPLACE"
 
@@ -422,7 +311,7 @@ def add_landscape(img: bpy.types.Image):
 
 	mod = act.modifiers.new(P_LAND_TEMP_NAME, "NODES")
 	
-	mod.node_group = get_or_make_displace_group(P_LAND_TEMP_NAME, image=img)
+	mod.node_group = nodes.get_or_make_displace_group(P_LAND_TEMP_NAME, image=img)
 
 	bpy.ops.object.mode_set(mode="OBJECT")	# modifiers can't be applied in EDIT mode
 
@@ -435,119 +324,13 @@ def add_landscape(img: bpy.types.Image):
 
 # -------------------------------------------------- Geometry Nodes
 
-def get_or_make_displace_group(name, image: bpy.types.Image=None):
-	if name in bpy.data.node_groups:
-		g = bpy.data.node_groups[name]
-		sockets = g.interface.items_tree
-
-		if not any(i for i in g.nodes if i.type == "IMAGE_TEXTURE" and i.name == "HYD_Displacement"):
-			n_image = g.nodes.new("GeometryNodeImageTexture")
-			n_image.label = "Displacement"
-			n_image.name = "HYD_Displacement"
-			n_image.extension = "EXTEND"
-			n_image.interpolation = "Cubic"
-			n_image.inputs[0].default_value = image
-			common.data.add_message(f"Existing group {name} was missing HYD_Displacement image node. It has been added, but hasn't been connected.", error=True)
-		elif not any(i for i in sockets if i.in_out == "OUTPUT" and i.socket_type == "NodeSocketGeometry") or\
-			not any(i for i in sockets if i.in_out == "INPUT" and i.socket_type == "NodeSocketGeometry"):
-			common.data.add_message(f"Updated existing group {name}, but it doesn't have Geometry input/output!", error=True)
-		else:
-			common.data.add_message(f"Updated existing group {name}.")
-		return g
-	else:
-		g = bpy.data.node_groups.new(name, type='GeometryNodeTree')
-		common.data.add_message(f"Created new group {name}.")
-
-		g.is_modifier = True
-		g.interface.new_socket("Geometry", in_out="INPUT", socket_type="NodeSocketGeometry")
-		i_scale = g.interface.new_socket("Scale", in_out="INPUT", socket_type="NodeSocketFloat")
-		i_scale.default_value = 1
-		i_scale.min_value = 0
-		i_scale.max_value = 2
-		i_scale.force_non_field = False
-		g.interface.new_socket("Displaced", in_out="OUTPUT", socket_type="NodeSocketGeometry")
-
-		nodes = g.nodes
-
-		n_input = nodes.new("NodeGroupInput")
-		n_output = nodes.new("NodeGroupOutput")
-
-		n_bounds = nodes.new("GeometryNodeBoundBox")
-		n_bounds.name = "HYD_Bounds"
-		n_pos = nodes.new("GeometryNodeInputPosition")
-		n_pos.name = "HYD_Position"
-
-		n_subpos = nodes.new("ShaderNodeVectorMath")
-		n_subpos.label = "Remove Offset"
-		n_subpos.name = "HYD_Get_Offset"
-		n_subpos.operation = "SUBTRACT"
-
-		n_subbound = nodes.new("ShaderNodeVectorMath")
-		n_subbound.label = "Width and Height"
-		n_subpos.name = "HYD_Get_Dimensions"
-		n_subbound.operation = "SUBTRACT"
-
-		n_normalize = nodes.new("ShaderNodeVectorMath")
-		n_normalize.label = "Normalize"
-		n_subpos.name = "HYD_Normalize"
-		n_normalize.operation = "DIVIDE"
-
-		n_image = nodes.new("GeometryNodeImageTexture")
-		n_image.label = "Displacement"
-		n_image.name = "HYD_Displacement"
-		n_image.extension = "EXTEND"
-		n_image.interpolation = "Cubic"
-		n_image.inputs[0].default_value = image
-
-		n_scale = nodes.new("ShaderNodeMath")
-		n_scale.label = "Scale"
-		n_scale.name = "HYD_Scale"
-		n_scale.operation = "MULTIPLY"
-		n_scale.inputs[1].default_value = 1
-
-		n_combine = nodes.new("ShaderNodeCombineXYZ")
-		n_combine.label = "Z Only"
-		n_combine.name = "HYD_Z_Only"
-		n_displace = nodes.new("GeometryNodeSetPosition")
-		n_displace.label = "Displace"
-
-		links = g.links
-
-		links.new(n_input.outputs["Geometry"], n_bounds.inputs[0])
-
-		links.new(n_pos.outputs[0], n_subpos.inputs[0])
-		links.new(n_bounds.outputs["Min"], n_subpos.inputs[1])
-
-		links.new(n_bounds.outputs["Max"], n_subbound.inputs[0])
-		links.new(n_bounds.outputs["Min"], n_subbound.inputs[1])
-
-		links.new(n_subpos.outputs[0], n_normalize.inputs[0])
-		links.new(n_subbound.outputs[0], n_normalize.inputs[1])
-
-		links.new(n_normalize.outputs[0], n_image.inputs["Vector"])
-
-		links.new(n_image.outputs["Color"], n_scale.inputs[0])
-		links.new(n_input.outputs["Scale"], n_scale.inputs[1])
-
-		links.new(n_scale.outputs[0], n_combine.inputs["Z"])
-
-		links.new(n_input.outputs["Geometry"], n_displace.inputs["Geometry"])
-		links.new(n_combine.outputs[0], n_displace.inputs["Offset"])
-
-		links.new(n_displace.outputs["Geometry"], n_output.inputs["Displaced"])
-
-		stagger_nodes(n_output, [n_displace], [n_combine], [n_scale], [n_image], [n_normalize], [n_subpos, n_subbound], [n_pos, n_bounds], [n_input], forwards=False)
-
-		return g
-
-def add_geometry_node(obj: bpy.types.Object, src: mgl.Texture):
+def add_geometry_nodes(obj: bpy.types.Object, img: bpy.types.Image):
 	"""Adds a Geometry Nodes modifier to the specified object.
 
 	:param obj: Object to add to.
 	:type obj: :class:`bpy.types.Object`
 	:param src: Source texture to add.
 	:type src: :class:`moderngl.Texture`"""
-	img = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
 
 	if len(obj.modifiers) == 0:
 		mod = obj.modifiers.new("HYD_" + obj.name, "NODES")
@@ -560,9 +343,9 @@ def add_geometry_node(obj: bpy.types.Object, src: mgl.Texture):
 	else:
 		mod = obj.modifiers[-1]
 	
-	mod.node_group = get_or_make_displace_group(f"HYD_{obj.name}", img)
+	mod.node_group = nodes.get_or_make_displace_group(f"HYD_{obj.name}", img)
 
-def add_into_geometry_nodes(obj: bpy.types.Object, src: mgl.Texture):
+def add_into_geometry_nodes(obj: bpy.types.Object, img: bpy.types.Image):
 	"""Adds the texture into an existing Geometry Nodes modifier.
 
 	:param obj: Object to add to.
@@ -570,7 +353,6 @@ def add_into_geometry_nodes(obj: bpy.types.Object, src: mgl.Texture):
 	:param src: Source texture to add.
 	:type src: :class:`moderngl.Texture`"""
 	data = common.data
-	img = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
 
 	mod = None
 	for m in obj.modifiers:
@@ -590,7 +372,7 @@ def add_into_geometry_nodes(obj: bpy.types.Object, src: mgl.Texture):
 	else:
 		output = outputs[0]
 
-	displace_group = get_or_make_displace_group("HYD_Displace")
+	displace_group = nodes.get_or_make_displace_group("HYD_Displace")
 
 	connected = None
 	if output.inputs[0].is_linked:
@@ -599,7 +381,7 @@ def add_into_geometry_nodes(obj: bpy.types.Object, src: mgl.Texture):
 		if connected.type == "GROUP" and connected.node_tree == displace_group:
 			connected.inputs[1].default_value = img
 			connected.inputs[2].default_value = 1
-			data.info += ["Updated existing displacement."]
+			data.add_message("Updated existing displacement.")
 			return
 	
 	g = nodes.new("GeometryNodeGroup")
@@ -612,13 +394,4 @@ def add_into_geometry_nodes(obj: bpy.types.Object, src: mgl.Texture):
 	if connected:
 		links.new(connected.outputs[0], g.inputs[0])
 
-	stagger_nodes(output, [g], forwards=True)
-
-def only_update(obj: bpy.types.Object, src: mgl.Texture):
-	"""Only writes the specified texture to a Blender Image used by other methods.
-
-	:param obj: Object to add to.
-	:type obj: :class:`bpy.types.Object`
-	:param src: Source texture to add.
-	:type src: :class:`moderngl.Texture`"""
-	_ = texture.write_image(f"HYD_{obj.name}_DISPLACE", src)
+	nodes.stagger_nodes(output, [g], forwards=True)
